requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
spacy>=3.7.4
